split news pages writing
