+++ modules/next/commands/helpers/makeNewsPage.ts
@@ -1,0 +1,50 @@
+import fs from "fs-extra";
+import path from "path";
+import prettier from "prettier";
+import { AlgolArticleData } from "../../../types";
+
+const newsFolder = path.join(__dirname, "../../pages/news");
+
+export const makeNewsPage = (data: AlgolArticleData) => {
+  const content = `
+  // Generated by the makeNewsPages command
+  
+  import React, { Fragment } from "react";
+  
+  import Head from "next/head";
+  import { ArticlePage } from "../../../../ui/src/components/ArticlePage";
+  import { pageActions } from "../../../helpers";
+  import { news } from "../../../../content/dist/news/${data.id}/news";
+  
+  export const Article = () => {
+    const title = \`${data.id} - ${data.title}\`;
+    const crumbs = [{ content: "News", url: "/news" }, { content: "${data.title}" }];
+    return (
+      <Fragment>
+        <Head>
+          <meta property="og:site_name" content="Chessicals"/>
+          <meta property="og:image" content="${data.mainImage}" />
+          <meta property="og:title" content={title} />
+          <meta property="og:description" content="${data.blurb}"/>
+          <title>{title}</title>
+        </Head>
+        <ArticlePage
+          crumbs={crumbs}
+          title={title}
+          actions={pageActions}
+          html={news}
+        />
+      </Fragment>
+    );
+  };
+  
+  export default Article;
+  `;
+  const out = path.join(newsFolder, data.slug);
+  fs.emptyDirSync(out);
+  fs.writeFileSync(
+    path.join(out, "index.tsx"),
+    prettier.format(content, { filepath: "foo.ts" })
+  );
+  console.log("Made news page for", data.title);
+};
