+++ modules/next/commands/makeNewsPages.ts
@@ -1,50 +1,19 @@
-import fs from "fs-extra";
-import path from "path";
-import prettier from "prettier";
+import { makeNewsPage } from "./helpers/makeNewsPage";
 import { newsList } from "../../content/dist/newsList";
 
-const out = path.join(__dirname, "../pages/news");
-fs.ensureDirSync(out);
+const newsId = process.argv[2];
 
-for (const data of newsList) {
-  const content = `
-// Generated by the makeNewsPages command
-
-import React, { Fragment, useEffect } from "react";
-
-import Head from "next/head";
-import { ArticlePage } from "../../../../ui/src/components/ArticlePage";
-import { pageActions } from "../../../helpers";
-import { news } from "../../../../content/dist/news/${data.id}/news";
-
-export const Article = () => {
-  const title = \`${data.id} - ${data.title}\`;
-  const crumbs = [{ content: "News", url: "/news" }, { content: "${data.title}" }];
-  return (
-    <Fragment>
-      <Head>
-        <meta property="og:site_name" content="Chessicals"/>
-        <meta property="og:image" content="${data.mainImage}" />
-        <meta property="og:title" content={title} />
-        <meta property="og:description" content="${data.blurb}"/>
-        <title>{title}</title>
-      </Head>
-      <ArticlePage
-        crumbs={crumbs}
-        title={title}
-        actions={pageActions}
-        html={news}
-      />
-    </Fragment>
-  );
-};
-
-export default Article;
-`;
-  fs.emptyDirSync(path.join(out, data.slug));
-  fs.writeFileSync(
-    path.join(out, data.slug, "index.tsx"),
-    prettier.format(content, { filepath: "foo.ts" })
-  );
-  console.log("Created page for news item", data.id, data.slug);
+if (!newsId) {
+  console.log("---- Making pages for all news ----");
+  for (const data of newsList) {
+    makeNewsPage(data);
+  }
+  console.log("---- All news got pages ----");
+} else {
+  const data = newsList.find(d => d.id === newsId);
+  if (!data) {
+    console.log(`News "${newsId}" doesn't exists!`);
+  } else {
+    makeNewsPage(data);
+  }
 }
