+++ modules/types/content.ts
@@ -7,7 +7,6 @@ export type AlgolArticleData = {
   created: string;
   updated: string;
   preloads: string[];
-  games: string[];
   thumbdata: string;
   mainImage: string;
 };
